@@ -1,12 +1,12 @@
 apply plugin: 'com.android.library'
 
 android {
-    compileSdkVersion 23
-    buildToolsVersion '26.0.2'
+    compileSdkVersion 28
+    buildToolsVersion '28.0.3'
 
     defaultConfig {
         minSdkVersion 14
-        targetSdkVersion 23
+        targetSdkVersion 28
         versionCode 9
         versionName "0.1.9"
 
@@ -24,10 +24,8 @@ android {
 }
 
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    testCompile 'junit:junit:4.12'
-    compile 'com.android.support:appcompat-v7:23.2.1'
-    compile 'com.daasuu:EasingInterpolator:1.0.0'
+    implementation "androidx.appcompat:appcompat:$androidx_version"
+    implementation "com.daasuu:EasingInterpolator:$ei_version"
 }
 
 //apply from: '../maven_push.gradle'
